requirements.txt
folium==0.14.0
haversine==2.8.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
plotly==5.15.0
streamlit==1.23.1
streamlit-folium==0.12.0